pyproject.toml
[tool.poetry]
name = "l2ss-py-autotest"
version = "0.1.0"
description = "Automated tests for new associations to l2ss-py service"
authors = ["PO.DAAC <podaac@jpl.nasa.gov>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "l2ss_py_autotest"}]

[tool.poetry.dependencies]
python = "^3.9"
python-cmr = "^0.8.0"
harmony-py = "^0.4.9"
matplotlib = "^3.7.2"
netCDF4 = "^1.6.4"
xarray = "^2023.6.0"
cf-xarray = "^0.8.3"
l2ss-py = "^2.4.0"
pygithub = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-html = "^3.2.0"
pytest-timeout = "^2.1.0"

[tool.pytest.ini_options]
junit_logging = "log"
log_cli = "True"
log_cli_level = "INFO"
log_level = "INFO"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cmr_association_diff = "l2ss_py_autotest.cmr_association_diff:run"
